+++ changelog.md
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+## Version 0.7
+
+### Additions:
+- accepts command line arguments; try `rpcalc --help`
+- accepts a limited stack size (command line `-s`)
+- accepts initial values (command line `-i`)
+- version can be checked via command line `--version`
+
 ## Version 0.6
 
 ### Fixes: